Change camelCase to kebab-case
