+++ plugins/wazuh_check_updates/public/components/current-update-details.tsx
