+++ plugins/wazuh_check_updates/public/components/dismiss-notification-check.tsx
