+++ plugins/wazuh_check_updates/public/components/updates-notification.tsx
