+++ plugins/wazuh_check_updates/public/hooks/available-updates.ts
