+++ plugins/wazuh_check_updates/public/hooks/user-preferences.ts
