+++ plugins/wazuh_check_updates/public/utils/get-current-available-update.ts
