+++ plugins/wazuh_check_updates/public/utils/index.ts
@@ -1,2 +1,2 @@
-export { getCurrentAvailableUpdate } from './getCurrentAvailableUpdate';
+export { getCurrentAvailableUpdate } from './get-current-available-update';
 export { formatUIDate } from './time';
