+++ plugins/wazuh_check_updates/server/routes/user-preferences/get-user-preferences.ts
