+++ plugins/wazuh_check_updates/server/routes/user-preferences/index.ts
@@ -1,6 +1,6 @@
 import { IRouter } from 'opensearch-dashboards/server';
-import { getUserPreferencesRoutes } from './getUserPreferences';
-import { updatePreferencesRoutes } from './updateUserPreferences';
+import { getUserPreferencesRoutes } from './get-user-preferences';
+import { updatePreferencesRoutes } from './update-user-preferences';
 
 export function userPreferencesRoutes(router: IRouter) {
   getUserPreferencesRoutes(router);
