+++ plugins/wazuh_check_updates/server/routes/user-preferences/update-user-preferences.ts
