+++ plugins/wazuh_check_updates/server/services/cronjob/job-scheduler-run.ts
