+++ plugins/wazuh_check_updates/server/services/savedObject/get-saved-object.ts
