+++ plugins/wazuh_check_updates/server/services/savedObject/set-saved-object.ts
