+++ plugins/wazuh_check_updates/server/services/settings/get-settings.ts
