+++ plugins/wazuh_check_updates/server/services/settings/update-settings.ts
