+++ plugins/wazuh_check_updates/server/services/updates/index.ts
@@ -1,1 +1,1 @@
-export { getUpdates } from './getUpdates';
+export { getUpdates } from './get-updates';
