+++ plugins/wazuh_check_updates/server/services/userPreferences/get-user-preferences.ts
