+++ plugins/wazuh_check_updates/server/services/userPreferences/update-user-preferences.ts
